+++ src/LibNoise.NET35/Filter/HybridMultiFractal.cs
@@ -68,17 +68,17 @@ namespace LibNoise.Filter
         /// <param name="x">The input coordinate on the x-axis.</param>
         /// <param name="y">The input coordinate on the y-axis.</param>
         /// <returns>The resulting output value.</returns>
-        public float GetValue(float x, float y)
+        public double GetValue(double x, double y)
         {
-            float signal;
+            double signal;
             int curOctave;
 
             x *= _frequency;
             y *= _frequency;
 
             // Initialize value : get first octave of function; later octaves are weighted
-            float value = _source2D.GetValue(x, y) + _offset;
-            float weight = _gain*value;
+            double value = _source2D.GetValue(x, y) + _offset;
+            double weight = _gain*value;
 
             x *= _lacunarity;
             y *= _lacunarity;
@@ -108,7 +108,7 @@ namespace LibNoise.Filter
             }
 
             //take care of remainder in _octaveCount
-            float remainder = _octaveCount - (int) _octaveCount;
+            double remainder = _octaveCount - (int) _octaveCount;
 
             if (remainder > 0.0f)
             {
@@ -132,9 +132,9 @@ namespace LibNoise.Filter
         /// <param name="y">The input coordinate on the y-axis.</param>
         /// <param name="z">The input coordinate on the z-axis.</param>
         /// <returns>The resulting output value.</returns>
-        public float GetValue(float x, float y, float z)
+        public double GetValue(double x, double y, double z)
         {
-            float signal;
+            double signal;
             int curOctave;
 
             x *= _frequency;
@@ -142,8 +142,8 @@ namespace LibNoise.Filter
             z *= _frequency;
 
             // Initialize value : get first octave of function; later octaves are weighted
-            float value = _source3D.GetValue(x, y, z) + _offset;
-            float weight = _gain*value;
+            double value = _source3D.GetValue(x, y, z) + _offset;
+            double weight = _gain*value;
 
             x *= _lacunarity;
             y *= _lacunarity;
@@ -175,7 +175,7 @@ namespace LibNoise.Filter
             }
 
             //take care of remainder in _octaveCount
-            float remainder = _octaveCount - (int) _octaveCount;
+            double remainder = _octaveCount - (int) _octaveCount;
 
             if (remainder > 0.0f)
             {
