+++ src/LibNoise.NET35/Filter/SumFractal.cs
@@ -44,7 +44,7 @@ namespace LibNoise.Filter
         /// <param name="x">The input coordinate on the x-axis.</param>
         /// <param name="y">The input coordinate on the y-axis.</param>
         /// <returns>The resulting output value.</returns>
-        public float GetValue(float x, float y)
+        public double GetValue(double x, double y)
         {
             int curOctave;
 
@@ -52,14 +52,14 @@ namespace LibNoise.Filter
             y *= _frequency;
 
             // Initialize value, fBM starts with 0
-            float value = 0;
+            double value = 0;
 
             // Inner loop of spectral construction, where the fractal is built
 
             for (curOctave = 0; curOctave < _octaveCount; curOctave++)
             {
                 // Get the coherent-noise value.
-                float signal = _source2D.GetValue(x, y)*_spectralWeights[curOctave];
+                double signal = _source2D.GetValue(x, y)*_spectralWeights[curOctave];
 
                 // Add the signal to the output value.
                 value += signal;
@@ -70,7 +70,7 @@ namespace LibNoise.Filter
             }
 
             //take care of remainder in _octaveCount
-            float remainder = _octaveCount - (int) _octaveCount;
+            double remainder = _octaveCount - (int) _octaveCount;
             if (remainder > 0.0f)
                 value += remainder*_source2D.GetValue(x, y)*_spectralWeights[curOctave];
 
@@ -88,10 +88,10 @@ namespace LibNoise.Filter
         /// <param name="y">The input coordinate on the y-axis.</param>
         /// <param name="z">The input coordinate on the z-axis.</param>
         /// <returns>The resulting output value.</returns>
-        public float GetValue(float x, float y, float z)
+        public double GetValue(double x, double y, double z)
         {
-            float signal;
-            float value;
+            double signal;
+            double value;
             int curOctave;
 
             x *= _frequency;
@@ -117,7 +117,7 @@ namespace LibNoise.Filter
             }
 
             //take care of remainder in _octaveCount
-            float remainder = _octaveCount - (int) _octaveCount;
+            double remainder = _octaveCount - (int) _octaveCount;
             if (remainder > 0.0f)
                 value += remainder*_source3D.GetValue(x, y, z)*_spectralWeights[curOctave];
