+++ src/LibNoise.NET35/Modifier/Invert.cs
@@ -45,7 +45,7 @@ namespace LibNoise.Modifier
         /// <param name="y">The input coordinate on the y-axis.</param>
         /// <param name="z">The input coordinate on the z-axis.</param>
         /// <returns>The resulting output value.</returns>
-        public float GetValue(float x, float y, float z)
+        public double GetValue(double x, double y, double z)
         {
             return -(((IModule3D) _sourceModule).GetValue(x, y, z));
         }
