+++ src/LibNoise.NET35/Primitive/Checkerboard.cs
@@ -42,7 +42,7 @@ namespace LibNoise.Primitive
         /// <param name="y">The input coordinate on the y-axis.</param>
         /// <param name="z">The input coordinate on the z-axis.</param>
         /// <returns>The resulting output value.</returns>
-        public float GetValue(float x, float y, float z)
+        public double GetValue(double x, double y, double z)
         {
             // Fast floor
             int ix = (x > 0.0 ? (int) x : (int) x - 1);
